Conflicts:
	README.md

+++ code/modules/paperwork/photocopier.dm
@@ -11,7 +11,7 @@
 	name = "photocopier"
 	desc = "Used to copy important documents and anatomy studies."
 	icon = 'icons/obj/library.dmi'
-	icon_state = "bigscanner"
+	icon_state = "photocopier"
 	anchored = 1
 	density = 1
 	use_power = 1
@@ -223,29 +223,32 @@
 /obj/machinery/photocopier/attackby(obj/item/O, mob/user, params)
 	if(istype(O, /obj/item/weapon/paper))
 		if(copier_empty())
-			user.drop_item()
+			if(!user.drop_item())
+				return
 			copy = O
 			O.loc = src
 			user << "<span class='notice'>You insert [O] into [src].</span>"
-			flick("bigscanner1", src)
+			flick("photocopier1", src)
 			updateUsrDialog()
 		else
 			user << "<span class='warning'>There is already something in [src]!</span>"
 
 	else if(istype(O, /obj/item/weapon/photo))
 		if(copier_empty())
-			user.drop_item()
+			if(!user.drop_item())
+				return
 			photocopy = O
 			O.loc = src
 			user << "<span class='notice'>You insert [O] into [src].</span>"
-			flick("bigscanner1", src)
+			flick("photocopier1", src)
 			updateUsrDialog()
 		else
 			user << "<span class='warning'>There is already something in [src]!</span>"
 
 	else if(istype(O, /obj/item/device/toner))
 		if(toner <= 0)
-			user.drop_item()
+			if(!user.drop_item())
+				return
 			qdel(O)
 			toner = 40
 			user << "<span class='notice'>You insert [O] into [src].</span>"
@@ -259,16 +262,11 @@
 			return
 		playsound(loc, 'sound/items/Ratchet.ogg', 50, 1)
 		user << "<span class='warning'>You start [anchored ? "unwrenching" : "wrenching"] [src]...</span>"
-		if(do_after(user, 20))
+		if(do_after(user, 20/O.toolspeed, target = src))
 			if(gc_destroyed)
 				return
 			user << "<span class='notice'>You [anchored ? "unwrench" : "wrench"] [src].</span>"
 			anchored = !anchored
-
-	else if(istype(O, /obj/item/weapon/grab)) //For ass-copying.
-		var/obj/item/weapon/grab/G = O
-		if(ismob(G.affecting) && G.affecting != ass)
-			MouseDrop_T(G.affecting, user)
 
 /obj/machinery/photocopier/ex_act(severity, target)
 	switch(severity)
@@ -306,7 +304,7 @@
 	else
 		user.visible_message("<span class='warning'>[user] starts putting [target] onto the photocopier!</span>", "<span class='notice'>You start putting [target] onto the photocopier...</span>")
 
-	if(do_after(user, 20))
+	if(do_after(user, 20, target = src))
 		if(!target || target.gc_destroyed || gc_destroyed || !Adjacent(target)) //check if the photocopier/target still exists.
 			return
